requirements.txt
customtkinter
opencv-python
pillow
pygame
faster-whisper
pyannote.audio==3.3.1
ffmpeg-python
reportlab
transformers
tensorboardX
numpy<2.0
pyannote.core<6.0
pyannote.metrics<4.0
pyannote.pipeline<4.0
